add userRoute and userService and implemented getAllUsers method
