+++ api/app.js
@@ -1,6 +1,7 @@
 import express from "express";
 import authRoute from "./routes/auth.route.js";
 import testRoute from "./routes/test.route.js";
+import userRoute from "./routes/user.route.js";
 import cookieParser from "cookie-parser";
 import {globalErrorHandler} from "./middleware/globalErrorHandler.js";
 
@@ -11,7 +12,7 @@ app.use(cookieParser())
 
 app.use("/api/v1/auth", authRoute)
 app.use("/api/v1/test", testRoute)
-
-app.use(globalErrorHandler())
+app.use("/api/v1/user", userRoute)
+app.use(globalErrorHandler());
 
 app.listen(8800)
