+++ api/controllers/auth.controller.js
@@ -1,11 +1,11 @@
-import * as userService from "../services/user.service.js";
+import * as authService from "../services/auth.service.js";
 
 export const register = async (req, res, next) => {
-  return await userService.register(req, res, next);
+  return await authService.register(req, res, next);
 };
 
 export const login = async (req, res, next) => {
-  return await userService.login(req, res, next);
+  return await authService.login(req, res, next);
 };
 
 export const logout = (req, res) => {
