+++ api/errors/UnexpectedError.js
@@ -1,0 +1,8 @@
+export class UnexpectedError extends Error {
+    constructor(message) {
+        if(message == ""){
+            message = "Unexpected error occurred!";
+        }
+        super(message);
+    }
+}
