+++ api/repositories/user.repository.js
@@ -21,3 +21,7 @@ export const addUser = async (user) => {
     data: user
   })
 }
+export const getAllUsers = async () => {
+  const users = await prisma.users.findMany();
+  return users;
+}
