+++ api/routes/user.route.js
@@ -1,0 +1,12 @@
+import express from "express";
+import { getUser, getUsers, updateUser, deleteUser } from "../controllers/user.controller.js";
+import { verifyToken } from "../middleware/verifyToken.js";
+
+const router = express.Router();
+
+router.get("/", getUsers);
+router.get("/:id", verifyToken, getUser);
+router.put("/:id", verifyToken, updateUser);
+router.delete("/:id", verifyToken, deleteUser);
+
+export default router;
